deployment/requirements_backend.txt
fastapi==0.115.7
numpy==2.2.2
onnxruntime==1.20.1
Pillow==11.1.0
uvicorn>=0.25.0
python-multipart
